backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
Pillow==10.4.0
transformers==4.44.0
torch==2.4.0
torchvision==0.19.0
ultralytics==8.2.0
httpx==0.27.0
python-dotenv==1.0.1
supabase==2.7.0
pydantic==2.9.0
pydantic-settings==2.5.0
